consolidated_project/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.2
pydantic-settings==2.6.1
anthropic==0.40.0
langgraph==0.2.53
langchain-anthropic==0.3.0
langchain-core==0.3.21
python-dotenv==1.0.1
httpx==0.28.0
sqlalchemy==2.0.36
asyncpg==0.30.0
greenlet==3.1.1
google-api-python-client==2.149.0
google-auth==2.36.0
